Send MQTT accept and delete notifications

+++ src/friendship/friendship.controller.ts
@@ -13,6 +13,7 @@ import SwaggerTags from '../common/swagger/tags/SwaggerTags.decorator';
 import { FriendRequestDto } from './dto/FriendRequest.dto';
 import { PlayerIdDto } from '../common/dto/player_id.dto';
 import FriendshipNotifier from './friendship.notifier';
+import { FriendshipDocument } from './friendship.schema';
 
 @SwaggerTags('Friendship')
 @Controller('friendship')
@@ -84,20 +85,34 @@ export class FriendshipController {
   @Post('accept/:_id')
   @UniformResponse(ModelName.FRIENDSHIP)
   async acceptRequest(@Param() param: _idDto, @LoggedUser() user: User) {
-    return await this.service.basicService.updateOne(
+    const filter = {
+      _id: param._id,
+      status: FriendshipStatus.PENDING,
+      requester: { $ne: user.player_id },
+      $or: [{ playerA: user.player_id }, { playerB: user.player_id }],
+    };
+    const [friendship, readError] =
+      await this.service.basicService.readOne<FriendshipDocument>({
+        filter,
+      });
+    if (readError) return [friendship, readError];
+
+    const result = await this.service.basicService.updateOne(
       {
         $set: { status: FriendshipStatus.ACCEPTED },
         $unset: { requester: '' },
       },
       {
-        filter: {
-          _id: param._id,
-          status: FriendshipStatus.PENDING,
-          requester: { $ne: user.player_id },
-          $or: [{ playerA: user.player_id }, { playerB: user.player_id }],
-        },
+        filter,
       },
     );
+    const [, updateError] = result;
+
+    if (!updateError) {
+      await this.notifier.friendRequestAccepted(friendship, user.player_id);
+    }
+
+    return result;
   }
 
   @ApiResponseDescription({
@@ -111,6 +126,15 @@ export class FriendshipController {
   @Delete(':_id')
   @UniformResponse(ModelName.FRIENDSHIP)
   async deleteFriendship(@Param() param: _idDto, @LoggedUser() user: User) {
+    const [friendship, readError] =
+      await this.service.basicService.readOne<FriendshipDocument>({
+        filter: {
+          _id: param._id,
+          $or: [{ playerA: user.player_id }, { playerB: user.player_id }],
+        },
+      });
+    if (readError) return readError;
+
     const [, error] = await this.service.basicService.deleteOne({
       filter: {
         _id: param._id,
@@ -118,5 +142,12 @@ export class FriendshipController {
       },
     });
     if (error) return error;
+
+    if (
+      friendship.status === FriendshipStatus.PENDING &&
+      friendship.requester.toString() !== user.player_id
+    ) {
+      await this.notifier.friendRequestRejected(friendship, user.player_id);
+    }
   }
 }
